requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
scikit-learn==0.24.2
pandas==1.3.0
numpy==1.21.0
python-dotenv==0.19.0
matplotlib==3.7.2
seaborn==0.12.2
werkzeug==2.3.7
flask-wtf==1.1.1
email-validator==2.0.0 